* 'feature/person_center' of https://github.com/zhiyicx/thinksns-plus-android: (109 commits)
  (#785)	code    fix  this
  (#813) code 我的提问-我的关注列表
  build  0.8.6.2   test
  (#817)   code   update  interface  to v2
  (#807) code 付费节点获取钱包余额
  (#816)  code   fix  this
  (#811)    code   fix  this
  ( #810)     code   fix  this
  (#809)    code     fix   this
  (#802) code fix bug
  update  net work error tip
  (#802) code fix bug
  (#808) code fix bug
  build v0.8.6.1 test
  (#801)  code   update  judgment
  (#805) code fix bug
  (#801) code fix bug
  (#801) code fix bug
  build 0.8.5.7
  ( #803)  cod   fix  this
  ...

# Conflicts:
#	app/src/main/java/com/zhiyicx/thinksnsplus/data/source/remote/UserInfoClient.java
#	app/src/main/java/com/zhiyicx/thinksnsplus/modules/personal_center/PersonalCenterFragment.java

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/home/message/messagereview/MessageReviewFragment.java
@@ -113,7 +113,7 @@ public class MessageReviewFragment extends TSListFragment<MessageReviewContract.
                 .with(getActivity())
                 .item1ClickListener(() -> {
                     mPresenter.approvedTopComment((long) topDynamicCommentBean.getFeed().getId(),
-                            topDynamicCommentBean.getComment().getId(), topDynamicCommentBean.getId().intValue());
+                            topDynamicCommentBean.getComment().getId().intValue(), topDynamicCommentBean.getId().intValue());
                     mReviewPopWindow.hide();
                 })
                 .item2ClickListener(() -> {
